Cave shaking
Added cave shaking to visually show the player something is happening. Additionally I wrote out all the text/info/lore that needs to be added to the in-game character journal providing the player with all the information needed to complete the puzzles. Just need to hook up the necessary components with the script I wrote. Finally I want to lock some information behind game events.